CS 302- Project 2
This program is similar to the classic ELIZA program first developed in
the 1960s but adapted to utilize specific concepts focused on in this
assignment. Eliza does basic natural language (human language)
processing. The program poses as a therapist and asks the user to type
in descriptions of their problems. The program will look for patterns of
words and reply with a prepared response based on the matched pattern.